% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kutnerALSM5e-package.R
\docType{package}
\name{kutnerALSM5e-package}
\alias{kutnerALSM5e}
\alias{kutnerALSM5e-package}
\title{A reproduction of the text book "Applied Linear Statistical Models 5th
Edition" by Kutner, Nachtsheim, Neter, and Li.}
\description{
Each chapter of the book is reproduced as a vignette with R code injected for
producing graphics, models, etc.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dewittpe/kutnerALST5e}
  \item \url{http://www.peteredewitt.com/kutnerALSM5e/}
}

}
\author{
\strong{Maintainer}: Peter DeWitt \email{dewittpe@gmail.com}

}
\keyword{internal}
