% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SENIC}
\alias{SENIC}
\title{SENIC}
\format{
a data.frame 
\tabular{clll}{
   \tab column name \tab Variable Name         \tab Description \cr
   \tab ----------- \tab -------------         \tab ----------- \cr
 1 \tab id          \tab Identification Number \tab 1 - 113 \cr
 2 \tab los         \tab Length of Stay        \tab Average Length of Stay of all \cr
   \tab             \tab                       \tab patients in hospital (in days) \cr
 3 \tab age         \tab Age                   \tab Average age of patients (in years)
 4 \tab             \tab                       \tab \cr
}
}
\usage{
SENIC
}
\description{
Study on the Efficany of Nosocomial Infection Control
}
\details{
See vignette("kutnerALSM5e-Appendix-C-Data-Sets") for details on this data
set's construciton.
}
\keyword{datasets}
