% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SENIC}
\alias{SENIC}
\title{SENIC}
\format{
a data.frame
}
\usage{
SENIC
}
\description{
Study on the Efficany of Nosocomial Infection Control.
}
\details{
See vignette("kutnerALSM5e-Appendix-C-Data-Sets") for details on this data
set's construciton.
}
\keyword{datasets}
